+++ src/ISettlerActions.sol
@@ -276,6 +276,8 @@ interface ISettlerActions {
         uint256 amountOutMin
     ) external;
 
+    function FLUXPOOL(address sellToken, uint256 ppm, bytes32 poolId, bool zeroForOne, uint256 minBuyAmount) external;
+
     function RENEGADE(
         address recipient,
         address sellToken,